Add some checks to make sure we're not running past the end of the RR we're dissecting.  (Some RR's still don't do that check, as they don't maintain an "rr_len" variable, and as "get_dns_name()" doesn't do any length checking.)
That obviates the need for having "compute_key_id()" check the length,
as we won't even get to calling it if the RR is too short.  Have it
still do a check for sanity, but just abort if the check fails, and have
it return -1 if the algorithm is unsupported.

Don't bother processing T_CERT RRs if we're not building a protocol
tree, don't fetch data from a T_CERT RR until we put it into the tree,
and fix the bug that change found, namely that we were using the wrong
variable in the "val_to_str()" call for the certificate type.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@9250 f5534014-38df-0310-8fa8-9805f1628bb7