The "cmd" members of "smb_saved_info_t" and "smb_info_t" structures always contain an SMB command code (SMB requests and responses both have command codes, so there's no "unknown" out-of-band value); make it a "guint8".
Make the argument to "decode_smb_name()" a "guint8" as an SMB command
code is passed to it ("guint8" and "unsigned char" are the same types on
all platforms we're likely to deal with, so it's a cosmetic change, not
a semantic one).

Put in an extra "GPOINTER_TO_UINT()" call before casting
"si->sip->extra_info" to "guint16", to squelch compiler warnings.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@9335 f5534014-38df-0310-8fa8-9805f1628bb7