Note in a comment that "dissect_asp_reply_get_status()" might be dissecting AFP server status - other servers might have different status formats.
In "dissect_asp_reply_get_status()", put the UTF-8 server name into a
tree, with the length and name in the tree as separate items, and fetch
the length into a 16-bit variable (as it's a 16-bit length in the
packet), as is done in "dissect_dsi_reply_get_status()".  (XXX - should
that just be done with an FT_UINT_STRING field, as is done for other
strings?)

Use "tvb_get_string()" to fetch the UTF-8 server name, and set the
length and name from the values we fetched, in both of those routines.

For FT_UINT_STRING fields in "dissect_asp_reply_get_status()" and
"dissect_dsi_reply_get_status()", don't fetch the length separately -
just use the value filled in by "proto_tree_add_item()" (now that a
"proto_item" is no longer opaque, we can do that).  That means we don't
have a problem with overflows of the 8-bit "len" variable if the length
is 255.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@9567 f5534014-38df-0310-8fa8-9805f1628bb7