Rename DISSECTOR_HELPER_{SRC,OBJECTS} to DISSECTOR_SUPPORT_{SRC,OBJECTS}.
Add some additional files, required by dissectors, to those lists.

Extract the stuff to get version information strings for libraries and
the OS, which is *not* needed by dissectors, from "util.c", which
contains routines that *are* needed by dissectors, and put it into a
separate file.

Make "dftest" link only with the dissector support stuff, not with all
of the Ethereal common files.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@9645 f5534014-38df-0310-8fa8-9805f1628bb7