Add a routine "retap_packet()" that runs through all packets, dissecting them and running all taps on them, but not reconstructing the packet list.  Use that in the IO-stat tap rather than "redissect_packet()"; the latter does more work and redraws the display, neither of which are necessary.
Call the filter callback when the Calc field is changed, to redraw the
graphs; that change also fixes things so that it's called when the
Filter field is changed.

Rename the "filter_button" member of an io_stat_graph_t to
"filter_field", as it's not the "Filter:" button, it's the text field
containing the filter expression.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@9659 f5534014-38df-0310-8fa8-9805f1628bb7