Add arguments to "dissect_xdlc_control()" to specify value_string tables, for use in the top-level item for the control field, for unnumbered frame function codes for commands and responses.  If the argument is null, default to the standard tables.
Use "dissect_xdlc_control()" and the #defines from "xdlc.h" for IrDA.

Use the reported length rather than the data length in the IrDA
dissector when iterating over the data in an IrDA packet.

Make "dissect_xid()" update the column information as well as the
protocol tree.  Put the slot number into the protocol tree only for
command frames - "the slot number field is undefined in discovery XID
response frames", to quote the IrLAP spec.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@9716 f5534014-38df-0310-8fa8-9805f1628bb7