Clean up indentation.
Make "font_zoom()" static - it's not used outside gtk/main.c.

Use "g_strdup()" rather than "strdup()", as we use "g_free()" to free
the result.

Put in a little more information when "font_zoom()" fails due to the
font name not being an XLFD font name, and don't continue if it fails.

Don't continue if the attempt to load the fonts fails, either.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@9813 f5534014-38df-0310-8fa8-9805f1628bb7