Check for write errors when writing CSV data as well.
Check for write errors when opening the CSV file to write the reverse
data information.

For some reason (perhaps having to do with the sort indicators in
titles, the column titles don't show up in the GtkCList structure, so
they don't get printed; just directly use the table of column titles.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@9832 f5534014-38df-0310-8fa8-9805f1628bb7