Cosmetic fix in WV-CSP tag definitions.
Flag a known VERY NASTY bug in WV-CSP 1.2 where a token is being defined
twice with the same binary identifier. It's already a mess that they don't
provide backwards compatible encodings; you can now write a WV-CSP document
without being able to parse it because of version issues (hint: 1.2 has no
public WBXML identifier yet and the spec explicitly asks to set it to
"Unknown/Missing" so you could write an 1.1 message in 1.2 format, which
looks *really* weird as the tokens are mixed in both codings).

Consequence: I have to foresee a discriminator for the WBXML dissectors
so I *may* predict the correct version. I like reviewing in engineering :)


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@10308 f5534014-38df-0310-8fa8-9805f1628bb7