Put in the URL of a page that discusses EIGRP packets.
Handle the prefix length and address similarly to the way that the BGP
dissector does, so that we mask out all the bits that the prefix length
tells us to, rather than just masking out entire bytes.

Check the prefix length so that we don't overflow the buffer for prefix
lengths > 32 - thanks to Stefan Esser for catching that one - and so
that we're not confused by a prefix length of 0.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@10324 f5534014-38df-0310-8fa8-9805f1628bb7