Redesigned the token map aggregation.
New simpler lists for both known registered WBXML public IDs and unknown
lists based on content type value. The latter have the possibility to add
a discriminator which facilitates a more correct/appropriate selection of
the token map in cases where more (backwards incompatible) versions exist.

Now the EXT_T tokens are correctly decoded, and a mapping function can be
assigned. This has been done in WML (existing code reused there) and in WV-CSP
(new functions looking up a value in a value_string based on the offset after
the EXT_T_0 token).

Add the textual WV-CSP media type to packet-text-media.c.

TODO - write the discriminators for WML, SyncML and WV-CSP. Add the code for
running the discriminators and possibly iterating over a list of mappings.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@10352 f5534014-38df-0310-8fa8-9805f1628bb7