Use a "while()" loop, rather than a "do { ... } while()" loop, in "dissect_rtps()", just in case a message with *no* submessages is sent.
Fetch the flags field in that loop, as we need the E bit value to
determine the byte order of the "next submessage offset" field; fetch
that value in the loop as well, and pass both those values, and the
byte-order flag, to submessage dissector routines as necessary.

Make the main protocol tree item for each submessage cover the entire
submessage; put into that protocol tree an item that covers the
submessage ID, with a named field.

Construct the submessage subtree in that loop, rather than in the
submessage dissectors.

Put into that subtree items for the flags and next submessage offset; we
do that in the submessage dissector, as the interpretation of the bits
in the flags field differs from submessage to submessage.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@10639 f5534014-38df-0310-8fa8-9805f1628bb7