DCERPC problem reported by JBM and identified by Todd Sabin
Other protocol, not only SMB will populate pinfo->private_data
thus   checking for existence of non-NULL pinfo->private_data is not sufficient
to determine   we have SMB data and this is what it is.

Refactor the adding of salt/FID from lower layer protocols and generalize it.
Create a new dissector_handle specific for SMB so that we know that IFF we came in through that handle, then whatever pinfo->private_data is what we expect it to be.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@11129 f5534014-38df-0310-8fa8-9805f1628bb7