The sequence number in COTP packets is either unused (for TP0) or relative to the beginning of the connection (for TP1-4).  They are *not* relative to the first segment of a segmented TPDU, so the sequence number can't be used in calls to "fragment_add_seq_check()".
For now, we assume that COTP packets are seen in order, and just use
"fragment_add_seq_next()" - that's unlikely to make anything worse (the
only time when "fragment_add_seq_check()" will work is when the first
segment has a sequence number of 0), and definitely fixes reassembly in
a number of cases.  In the long term, we might want to handle
out-of-order packets above the reassembly layer, so that it handles even
out-of-order unsegmented packets.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@11188 f5534014-38df-0310-8fa8-9805f1628bb7