Make the elapsed time value an "unsigned int" (if the stop time is before the start time, due to packet time stamp bogosity, neither a signed value nor an unsigned value will give a sensible answer, and it doesn't need to be a "long" as the capture's unlikely to span more than 2^32-1 seconds, again modulo packet time stamp bogosity).  That eliminates some format item vs. value warnings.
Make the packet count an "unsigned int" (it can't be negative, and we're
unlikely to have more than 2^32-1 packets - and, if we are, it should be
made a guint64).  That eliminates some more format item vs. value
warnings.

Initialize "start_time" and "stop_time" to 0, so that we handle files
with no packets (and so that we don't get uninitialized variable
warnings).

Display the file size with PRIu64, after casting it to "guint64" (that
way we handle file sizes if they're "unsigned int", "unsigned long", or
a 64-bit type on ILP64 platforms (it's 64 bits in BSDs derived from
4.4BSD, even on ILP64 platforms).

Mark the first argument to "file_open_entry_changed()" as unused.

Don't use G_CONST_RETURN, as it's not defined by GLib 1.2[.x].

Get rid of an unused variable.

Clean up white space.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@11189 f5534014-38df-0310-8fa8-9805f1628bb7