remove two preference options related to iscsi header digest
make ethereal attempt to automatically detect wether header digest is used or not for iscsi sessions.

This makes ethereal decode the packets properly EVEN for perfectly normal sessions where
the discovery session is performed with no digest but the normal login session negotiates digest.


the detected headerdigest setting is tcp session wide and thus it
it does not work for such initiators (if such exist) that resuse the same socketpair between the discovery and normal login sessions.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@11850 f5534014-38df-0310-8fa8-9805f1628bb7