Split the preview function into a prepare and a do part, to be able to do some experiments about the preview timeout things.
As the FileChooser dialog (GTK2.4) doesn't seem to let pass other GTK events (even with calling gtk_main_iteration()), it's not possible to change the preview behaviour to be more keen. To get a mechanism to update the preview independant from the dialog workings, we would need to start a new thread/task, which is too much effort for this IMHO.

So we might use the more simple approach of setting the timeout by a preference setting.

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@11856 f5534014-38df-0310-8fa8-9805f1628bb7