Preference strings don't remain null pointers for long - they get converted to pointers to null strings (see prefs.c for an explanation). Initialize "current_asn1" to a mallocated null string.
That lets us use "strcmp()" to compare asn1_filename and current_asn1. 
We already do that with asn1_pduname and current_pduname, so use
"strcmp()" for them as well.

Call "build_pdu_tree()" if we haven't yet built a PDU tree, even if the
PDU type preference hasn't changed.

Don't register with any of the ports unless we have a PDU tree, so we
don't try to dissect with a null PDU tree pointer.

Before constructing a PDU tree, if we already have a PDU tree, free it.

When we free data_nodes, free the data attached to the nodes.

Make a bunch of stuff static.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@12045 f5534014-38df-0310-8fa8-9805f1628bb7