From Chris Maynard: free strings fetched with "tvb_get_string()".
Use "format_text()" or "tvb_format_text()" with the %s format, to handle
non-printable characters.

When uncompressing data into a new tvbuff, add it to the list of data
sources as we do elsewhere.

When putting items into the protocol tree from an uncompressed tvbuff,
use the offset into the tvbuff and the length of the item.

Fetch string lengths into a variable and use the variable rather than
repeatedly fetching the length from a tvbuff.

Use -1 rather than tvb_length_remaining() when putting a "to the end of
the tvbuff" item into the protocol tree.

Use "proto_tree_add_item()" to put items into the tree, rather than
"proto_tree_add_xxx" calls with a "tvb_get_yyy" call as the argument.

If we've fetched a value into a variable, don't re-fetch it in
"proto_tree_add_xxx" calls.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@12971 f5534014-38df-0310-8fa8-9805f1628bb7