Rename the FieldError exception to DissectorError.
Add a DISSECTOR_ASSERT() macro, which is the usual type of assertion
macro, but throws a DissectorError exception with a message giving the
flien and line number and the failed test as a string.  Use that macro
in "alloc_field_info()".

Report that exception in the Info column and the protocol tree, as well
as logging the exception failure with g_warning().


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@13078 f5534014-38df-0310-8fa8-9805f1628bb7