there are two NETLOGON protocols
the really really old one which is never used
and there is the DCE/RPC one that is used by all windows/cifs implementations.

We used to reserve the protocol name NETLOGON for the old obsolete protocol
and since we can not have two protocols with the same name, called
the current protocol RPC_NETLOGON instead.

that does not exactly make sense.


This renames the old obsolete protocol to SMB_NETLOGON  to make way and allow
the real netlogon protocol take the name NETLOGON




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@13215 f5534014-38df-0310-8fa8-9805f1628bb7