Capitalize "Net-SNMP" the way the Net-SNMP project capitalizes it.
Note that the "=PATH" argument to "--with-net-snmp" is optional.

If "--with-net-snmp" is specified, give an error if Net-SNMP is found
but "--with-ssl" wasn't specified, as the user has specified that they
really want a version of Ethereal with Net-SNMP; if it's not specified,
just warn of that and drive on, as we did before.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@13786 f5534014-38df-0310-8fa8-9805f1628bb7