Fix the display of help information for various --with arguments (for whatever reason, the changequote() hack doesn't seem to be working, at least on OS X; the square brackets were simply being removed from the messages).
Rename the "ifpresent" value for the --with variables to "ifavailable",
to match the help messages.

At least for the SNMP libraries, have the "checking" message just say
"whether to use XXX" and have the result message say "yes, if available"
if the user didn't explicitly specify "--with" or "--without", as the
default just leaves the library out if it doesn't work, but explicitly
specifying "--with" means you get an error if it doesn't work.

Make "--with-net-snmp" fail if the headers aren't present (if the user
explicitly asks for something, we should fail if we can't give it to
them, as they presumably really wanted it).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@13789 f5534014-38df-0310-8fa8-9805f1628bb7