add the missing MODE field in mover_get_state that was introduced in version 4.
also change bytes_left_to_read to be a INT64 and not a UINT64.
It iss supposed to be an UINT64  but some implementations are buggy and it is easier to make sense of this field one for those buggy clients if it is printed as an INT64.
This change/violation will not affect any correct/valid NDMP implementation or how ethereal decodes its output  unless your entire backup is > 2**63 bytes in size.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@14228 f5534014-38df-0310-8fa8-9805f1628bb7