bugfix: in an attepmt to fix a crash while doing reassemling of connection oriented packets, some weeks ago I've added some tvb_ensure_bytes_exist() calls to test the case that the alloc_hint value is bogus (found in private fuzz testing).
Unfortunately, this is *very* safe now as it will mark a lot of DCE/RPC packets as malformed, thus the reassembling code won't work with most packets :-(

I've replaced this with the correct check and used THROW(ReportedBoundsError) instead which is hopefully the right error in this case.

BTW: could someone please add a description of the ReportedBoundsError vs. BoundsError to the code as I still don't get it.

In addition: Use a toplevel tree item to display the DCE/RPC defragmentation just as it's done in the TCP desegmentation code.

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@15188 f5534014-38df-0310-8fa8-9805f1628bb7