Add a "register_dfilter_stat()", to register stats that take a display filter as an argument on the command line and have a dialog box to enter the display filter through the GUI.  Use it for all stats using "gtk_tap_dfilter_dlg_cb()".
Add a top-level "stat_menu.h" file to declare "REGISTER_STAT_GROUP_E"
for the benefit of the declaration of "register_dfilter_stat()" in the
top-level "tap_dfilter_dlg.h".  Rename the "stat_menu.h" in the gtk
directory to "gtk_stat_menu.h", so as not to have two headers with the
same name.

Get rid of headers not declaring any functions not being used in the
module.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@15493 f5534014-38df-0310-8fa8-9805f1628bb7