When jumping (e.g. using menu item Go/Go to Packet...) to a new packet which is currently not displayed, this newly selected packet will become the first in the packet list.
This is inconvenient, as most of the time (at least) I'm interested not only in the packets behind the newly selected one, but also to have some history *before* it.

So this change will scroll the packet list to have the selected packet after the first third of the packet list.

This change won't take effect if the new packet is already visible (only the selection is changed) or it's near the beginning or end of the packet list (so the whole beginning/end of the list is shown).

git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@15772 f5534014-38df-0310-8fa8-9805f1628bb7