A loop processing all the bytes in a tvbuff from a given offset to the end of the data in the tvbuff should stop when the offset is >= the total amount of data in the tvbuff, not when it's > the total amount in the tvbuff following the starting offset.
In "unpack_digits()", return a null string, not a null pointer, if
there's nothing left in the tvbuff starting at the starting offset, so
that the caller doesn't have to check for a null pointer, and return an
ep_alloc()ed buffer, so the caller doesn't have to worry about freeing
the result.

If we see a filler digit, don't advance the offset in the string buffer;
we want to put the terminating '\0' right after the character we just
put into the string.

Fuzzed against some GSM captures.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@16002 f5534014-38df-0310-8fa8-9805f1628bb7