Give more details about too-short AVPs, give up immediately for AVPs shorter than 2 bytes, and make the item for an AVP with a length < 2 a generated item.
Put the top-level item for an AVP into the tree the same way regardless
of whether it's Vendor-Specific or not, and skip past the type and
length right after that, before we check for Vendor-Specific.  (This
means we no longer treat "vendor ID = 0" as an indication that this
isn't Vendor-Specific - nothing prevents a packet from getting onto the
wire with a vendor ID of 0; this fixes bug 485.)

Don't require a Vendor-Specific AVP to be at least 6 bytes long; it
might not be particularly useful to have one that has a vendor ID and
nothing else, but we might as well dissect the vendor ID portion.

Do some other cleanups.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@16015 f5534014-38df-0310-8fa8-9805f1628bb7