in svn 15335   the tcp analysis was changed to do its stuff and to populate (prepend to) COL_INFO before callking the subdissectors instead of calling the tcp analysis (and prepend colingo) eitehr after the subdissector returned normally  or if an exception caused by a subdissector was rised.
this as a sideffect caused tcp analysis data to be overwritten if the subdissector caused any output to the info column. (and made tcp analysis suboptimal)


this change adds a new function   col_prepend_fence_fstr()  that will prepend
the info column with the string and also, if there was no fence already defined, create a fence and set it after the prepended col info text.

This way, even if the subdissectors generate and rewrite col info, the tcp analysis data will still be displayed on the info column.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@16116 f5534014-38df-0310-8fa8-9805f1628bb7