Add #defines for NDS tag values, and use them in the definition of nds_tags[] and in switch statements.  (Also, catch a case where we didn't use NDS_PTYPE_ values.)
Use the existing #defines for MVTYPE_ values.

Make the handling of NDS_PTYPE_{IP,UDP,TCP} in one case match the
handling of the other cases (don't advance "ioffset", as it's advanced
by the length of the item).

When checking whether an NDS reply has a completion code, don't check
the fragment size, check the reassembled packet size.

Fix the handling of referrals in one case.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@16291 f5534014-38df-0310-8fa8-9805f1628bb7