Give RFC numbers for MIPv6, NEMO, and FMIPv6.
Fix up NEMO #defines for field offsets, and field and option lengths, to
match #defines for other options.

Handle the padding in the FMIPv6 LLA option internally to the option
parser, and don't include the link-layer address unless it's present. 
Don't handle the length of the LLA option specially in the generic
option parser.

Clean up white space.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@16361 f5534014-38df-0310-8fa8-9805f1628bb7