some ACEs are not used for authorization and thus do NOT contain an access mask and a SID.
make the dissection of the ACL check the type for each individual ACE and only dissect  as access mask and sid  those ACEs we know how to handle.


this prevents ethereal from dumping on w32 if we encounter any of these "special" ACE entries,   such as the ones used for storing location data for offline files.





git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@16881 f5534014-38df-0310-8fa8-9805f1628bb7