Jon Oberheide:
Ethereal does not take into account the protocol field of the IP header
when reassembling fragmented packets as specified in RFC791.  This can
lead to incorrect reassembly of packets with an identical src address,
dst address, and identification number, but with differing protocols.

The attached patch includes the protocol in the generation of the id
used to index into the reassembly table.



git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@16937 f5534014-38df-0310-8fa8-9805f1628bb7