Distinguish between "writing to the standard output" and "writing to a pipe"; there's not much point in writing to the standard output if you're *not* writing to a pipe, but....
"-b" doesn't necessarily imply a ring buffer - you can just request that
Tethereal keep switching files forever.

Standardize on an exit status of 1 for all those errors (there's a
sort-of convention, adhered to by many apps, that an exit status of 1
means a command-line argument error (as in "illegal flag" or "you
combined two flags that don't make sense together") and an exit status
of 2 is for other "run-time" errors.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@16942 f5534014-38df-0310-8fa8-9805f1628bb7