updates to get the fc and scsi dissectors and get rid of some breakage in the design
let the scsi transport keep track of itl (initiator, target, lun) matching
and let it pass a itl structure to scsi   that is persistent across packets.

let scsi use this itl structure to track device type for a specific itl instead of the (must have been) broken hashtable.

update both iscsi and fc to track the itl structure for scsi and schange the scsi signature to accept itl as a parameter.


more to come.




git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@17942 f5534014-38df-0310-8fa8-9805f1628bb7